Bankruptcy function broken, upgrade functionality works....quite well
while it will remove all of the improvements perfectly, it will not
change ownership of the property.  I won't get back around to this until
much later (if someone else wants to fix it, go ahead)

as far as upgrade(), I've tested it quite extensively, but I could've
missed something.  It will only show: properties where the user owns all
of Color-sibling-properties, properties where there is room for
improvement, & properties where the user can afford to buy an upgrade.d