added drawBoard()
Squares are a bit on the small size so i had to cut the string names to 5 characters at most to accommodate for the minimum size from pre-condition on my drawBoard(). Feel free to change if you find a better way to make string length change for re-sizing the board for bigger sizes.